requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
httpx==0.27.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
pydantic==2.9.2
coinbase-advanced-py==1.3.0
web3==6.15.0
validators==0.22.0
aiohttp==3.9.1
google-generativeai==0.8.3
ddgs==0.2.2
